streamlit/report/requirements.txt
pandas==2.0.3
streamlit==1.11.1
plotly==4.14.3
mxnet-mkl==1.6.0
numpy
matplotlib==3.6.3
seaborn==0.12.2
scikit-learn==1.2.1
Pillow==9.4.0
wordcloud==1.8.2.2
altair==4.0
st-pages==0.4.1